Optimized & cleaned up travis CI builds and deployment

+++ deploy.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# Script used for travis to push build folder content to gh-pages branch
+# Script used by travis to push build folder content to gh-pages branch
 
 # Get repo url (should be https://github.com/user/repo.git)
 url=$(git config remote.origin.url)
@@ -15,8 +15,8 @@ user="${split[1]}"
 repo="${split[2]}"
 
 # Clone master branch from user repo
-git clone --quiet "https://$user:${GH_TOKEN}@github.com/$user/$repo.git" --branch=master gh-pages
-cd gh-pages
+git clone --quiet "https://$user:${GH_TOKEN}@github.com/$user/$repo.git" --branch=master source
+cd source
 
 # Get latest commit ID from master branch
 head=$(git log --format="%h" -n 1)
@@ -25,12 +25,15 @@ head=$(git log --format="%h" -n 1)
 git checkout --quiet gh-pages
 cp -rf ../build/* .
 git add -A
-git status
 
-# Setup travis git user
-git config user.name "travis"
-git config user.email "travis@email.com"
+# Check for changes
+status=$(git status)
+echo "$status";
 
-# Commit and push
-git commit -m "CI Deployment to Github Pages ($user@$head)"
-git push --force --quiet "https://${GH_TOKEN}@$remote" gh-pages:gh-pages > /dev/null 2>&1
+# Setup travis git user, commit and push changes
+if [[ $status != *"nothing to commit"* ]] ; then
+  git config user.name "travis"
+  git config user.email "travis@email.com" 
+  git commit -m "CI Deployment to Github Pages ($user@$head)"
+  git push --force --quiet "https://${GH_TOKEN}@$remote" gh-pages:gh-pages > /dev/null 2>&1
+fi
